+++ expyriment/stimuli/_audio.py
@@ -10,12 +10,11 @@ Oliver Lindemann <oliver@expyriment.org>'
 
 import os
 
-try:
-    import android.mixer as mixer
-except Exception:
-    import pygame.mixer as mixer
+import pygame
+
 from .. import _internals
 from ..misc import unicode2byte
+from ..misc._timer import get_time
 from ._stimulus import Stimulus
 
 
@@ -37,11 +36,9 @@ class Audio(Stimulus):
         Parameters
         ----------
         filename : str
-            the filename. Must be an .ogg or uncompressed .wav file.
+            filename (incl. path) of the audio file
 
         """
-        if os.path.splitext(filename)[1] not in ('.wav', '.ogg'):
-            raise ValueError("The audio file must be an .ogg or uncompressed .wav file")
 
         Stimulus.__init__(self, filename)
         self._filename = filename
@@ -93,15 +90,25 @@ class Audio(Stimulus):
         return rtn
 
     def preload(self):
-        """Preload stimulus to memory."""
+        """Preload stimulus to memory.
 
+        Returns
+        -------
+        time : int
+            the time it took to execute this method
+
+        """
+
+        start = get_time()
         if not self._is_preloaded:
-        # Due to a bug in handling file names in PyGame 1.9.2, we pass a file
-        # handle to PyGame. See also:
-        # https://github.com/expyriment/expyriment/issues/81
+            # Due to a bug in handling file names introduced in PyGame 1.9.2,
+            # we pass a file handle to PyGame. See also:
+            # https://github.com/expyriment/expyriment/issues/81
             with open(self._filename, 'rb') as f:
-                self._file = mixer.Sound(f)
+                self._file = pygame.mixer.Sound(f)
             self._is_preloaded = True
+
+        return int((get_time() - start) * 1000)
 
     def unload(self, **kwargs):
         """Unload stimulus from memory.
@@ -109,11 +116,19 @@ class Audio(Stimulus):
         This removes the reference to the object in memory.
         It is up to the garbage collector to actually remove it from memory.
 
+        Returns
+        -------
+        time : int
+            the time it took to execute this method
+
         """
 
+        start = get_time()
         if self._is_preloaded:
             self._file = None
             self._is_preloaded = False
+
+        return int((get_time() - start) * 1000)
 
     def play(self, loops=0, maxtime=0, fade_ms=0, log_event_tag=None):
         """Play the audio stimulus.