+++ expyriment/stimuli/_visual.py
@@ -205,6 +205,7 @@ class Visual(Stimulus):
             width, height = len(surf[0]), len(surf)
         ogl.glEnable(ogl.GL_TEXTURE_2D)
         ogl.glBindTexture(ogl.GL_TEXTURE_2D, txtr)
+        ogl.glPixelStorei(ogl.GL_UNPACK_ALIGNMENT, 1)
         ogl.glTexImage2D(ogl.GL_TEXTURE_2D, 0, colours, width, height, 0,
           colours, ogl.GL_UNSIGNED_BYTE, textureData)
         ogl.glTexParameterf(ogl.GL_TEXTURE_2D,