+++ expyriment/stimuli/defaults.py
@@ -100,8 +100,9 @@ rectangle_corner_anti_aliasing = 0
 picture_position = (0, 0)
 
 # Video
+video_resizing = (None, None)
+video_audio_backend = "pygame"
 video_position = [0, 0]
-video_backend = "mediadecoder"  # 'mediadecoder' or 'pygame'
 
 # Tone
 tone_frequency = 440